# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Timeful.",
        "title": "Timeful API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3002/api",
    "paths": {
...
        "/analytics/user/{email}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Gets the user by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/events/{eventId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Gets an event based on its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Browser Event Visitor Identity public ID",
                        "name": "eventVisitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns server-proven owner capabilities and browser eventVisitorId; response entries add publicId and canEdit.",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Event"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "canCreateResponse": {
                                            "type": "boolean"
                                        },
                                        "canEditSettings": {
                                            "type": "boolean"
                                        },
                                        "canManageEvent": {
                                            "type": "boolean"
                                        },
                                        "eventVisitorId": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Requires Event Owner Edit Token proof, the associated Platform Visitor Identity session, or an owner-issued Granted EVCC; base EVCCs never authorize settings edits. Archived events are read-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Edits an event based on its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Timed events require the complete canonical slot contract; day-only events require dates",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "activeSlots": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "attendees": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "blindAvailabilityEnabled": {
                                    "type": "boolean"
                                },
                                "dates": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "daysOnly": {
                                    "type": "boolean"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "eventTimezone": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "notificationsEnabled": {
                                    "type": "boolean"
                                },
                                "remindees": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "sendEmailAfterXResponses": {
                                    "type": "integer"
                                },
                                "signUpBlocks": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.SignUpBlock"
                                    }
                                },
                                "slotGeneration": {
                                    "$ref": "#/definitions/models.SlotGeneration"
                                },
                                "timedRecurrence": {
                                    "$ref": "#/definitions/models.TimedRecurrence"
                                },
                                "type": {
                                    "$ref": "#/definitions/models.EventType"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Event name must be 100 characters or fewer",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "403": {
                        "description": "Owner authority required or event archived",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Requires the same owner credentials as settings edits; deleted events and responses stop resolving.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Deletes an event based on its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Owner authority required or event archived",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/events/{eventId}/archive": {
            "post": {
                "description": "Requires the same owner credentials as settings edits; archive makes the event read-only and unarchive restores mutations.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Archive an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Archive status",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "archive": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Owner authority required or event archived",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
        "/users/{userId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Returns a minimal public user profile (safe for unauthenticated clients)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.AppleCalendarAuth": {
            "type": "object"
        },
...
        "models.CalendarAccount": {
            "type": "object",
            "properties": {
                "appleCalendarAuth": {
                    "$ref": "#/definitions/models.AppleCalendarAuth"
                },
                "calendarType": {
                    "$ref": "#/definitions/models.CalendarType"
                },
                "email": {
                    "description": "Email is required for all calendar accounts",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "icsCalendarAuth": {
                    "$ref": "#/definitions/models.ICSCalendarAuth"
                },
                "oAuth2CalendarAuth": {
                    "$ref": "#/definitions/models.OAuth2CalendarAuth"
                },
                "picture": {
                    "type": "string"
                },
                "subCalendars": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SubCalendar"
                    }
                }
            }
        },
        "models.CalendarEvent": {
            "type": "object",
            "properties": {
                "allDay": {
                    "description": "Whether the event is an all day event",
                    "type": "boolean"
                },
                "calendarId": {
                    "type": "string"
                },
                "endDate": {
                    "type": "integer"
                },
                "free": {
                    "description": "Whether the user is free during this event",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "startDate": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "models.CalendarOptions": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "$ref": "#/definitions/models.BufferTimeOptions"
                },
                "workingHours": {
                    "$ref": "#/definitions/models.WorkingHoursOptions"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "activeSlots": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "blindAvailabilityEnabled": {
                    "description": "Whether to enable blind availability",
                    "type": "boolean"
                },
                "calendarEventId": {
                    "type": "string"
                },
                "collectEmails": {
                    "type": "boolean"
                },
                "creatorPosthogId": {
                    "description": "PostHog ID for the event creator",
                    "type": "string"
                },
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "daysOnly": {
                    "description": "Whether to only poll for days, not times",
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "eventTimezone": {
                    "type": "string"
                },
                "hasResponded": {
                    "description": "Whether the current viewer has responded to the availability group",
                    "type": "boolean"
                },
                "hasSpecificTimes": {
                    "description": "Used for specific times for specific dates feature",
                    "type": "boolean"
                },
                "isArchived": {
                    "type": "boolean"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isSignUpForm": {
                    "description": "Sign up form details",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "notificationsEnabled": {
                    "type": "boolean"
                },
                "numResponses": {
                    "description": "Used to store the number of responses for the event",
                    "type": "integer"
                },
                "ownerId": {
                    "type": "string"
                },
                "remindees": {
                    "description": "Remindees",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Remindee"
                    }
                },
                "responses": {
                    "description": "Availability responses",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.Response"
                    }
                },
                "scheduledEvent": {
                    "description": "Scheduled event",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CalendarEvent"
                        }
                    ]
                },
                "sendEmailAfterXResponses": {
                    "type": "integer"
                },
                "shortId": {
                    "type": "string"
                },
                "signUpBlocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SignUpBlock"
                    }
                },
                "signUpResponses": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.SignUpResponse"
                    }
                },
                "slotGeneration": {
                    "$ref": "#/definitions/models.SlotGeneration"
                },
                "startOnMonday": {
                    "description": "Whether to start the event on Monday (as opposed to Sunday, used for DOW events)",
                    "type": "boolean"
                },
                "timeIncrement": {
                    "type": "integer"
                },
                "timedRecurrence": {
                    "$ref": "#/definitions/models.TimedRecurrence"
                },
                "times": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "$ref": "#/definitions/models.EventType"
                },
                "when2meetHref": {
                    "type": "string"
                }
            }
        },
        "models.EventType": {
            "type": "string",
            "enum": [
                "specific_dates",
                "dow",
                "group"
            ],
            "x-enum-varnames": [
                "SPECIFIC_DATES",
                "DOW",
                "GROUP"
            ]
        },
...
        "models.OAuth2CalendarAuth": {
            "type": "object"
        },
        "models.Remindee": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "responded": {
                    "type": "boolean"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "availability": {
                    "description": "Availability",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "calendarOptions": {
                    "$ref": "#/definitions/models.CalendarOptions"
                },
                "email": {
                    "type": "string"
                },
                "enabledCalendars": {
                    "description": "Maps email to an array of sub calendar ids",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "guestEditPolicy": {
                    "type": "string"
                },
                "guestId": {
                    "type": "string"
                },
                "guestOwnershipMode": {
                    "type": "string"
                },
                "ifNeeded": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "manualAvailability": {
                    "description": "Mapping from the start date of a day to the available times for that day",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                },
                "name": {
                    "description": "Guest information",
                    "type": "string"
                },
                "useCalendarAvailability": {
                    "description": "Calendar availability variables for Availability Groups feature",
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userId": {
                    "description": "User information",
                    "type": "string"
                }
            }
        },
        "models.SignUpBlock": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Id is a client-provided opaque block identity. The server owns block\nidentities in event_signup_blocks: only a canonical UUID that names an\nexisting block on the event keeps that identity, and every other value is\nignored on write.",
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "startDate": {
                    "type": "integer"
                }
            }
        },
        "models.SignUpResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "description": "Guest information",
                    "type": "string"
                },
                "signUpBlockIds": {
                    "description": "The IDs of the sign up blocks that the user has signed up for",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userId": {
                    "description": "User information",
                    "type": "string"
                }
            }
        },
        "models.SlotGeneration": {
            "type": "object",
            "properties": {
                "endTimeLocal": {
                    "type": "string"
                },
                "startTimeLocal": {
                    "type": "string"
                },
                "timeIncrementMinutes": {
                    "type": "integer"
                }
            }
        },
...
        "models.TimedRecurrence": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "selectedDays": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "selectedDaysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "startOnMonday": {
                    "type": "boolean"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Profile info",
                    "type": "string"
                },
                "calendarAccounts": {
                    "description": "CalendarAccounts is a mapping from {`email_CALENDARTYPE` =\u003e CalendarAccount} that contains all the\nadditional accounts the user wants to see google calendar events for",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.CalendarAccount"
                    }
                },
                "calendarOptions": {
                    "description": "Calendar options",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CalendarOptions"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasCustomName": {
                    "description": "Whether the user has set a custom name for themselves, i.e. don't change their name when they sign in",
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numEventsCreated": {
                    "type": "integer"
                },
                "picture": {
                    "type": "string"
                },
                "primaryAccountKey": {
                    "description": "The calendarAccountKey of the account the user first signed in with",
                    "type": "string"
                },
                "timezoneOffset": {
                    "type": "integer"
                }
            }
        },
...
        "responses.Error": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {}
            }
        },
...
    }
}